+++ rust/scx_utils/Cargo.toml
@@ -19,7 +19,7 @@ lazy_static = "1.4"
 libbpf-cargo = "0.24.1"
 libbpf-rs = "0.24.1"
 log = "0.4.17"
-nvml-wrapper = "0.10.0"
+nvml-wrapper = { version = "0.10.0", optional = true }
 paste = "1.0"
 regex = "1.10"
 scx_stats = { path = "../scx_stats", version = "1.0.4" }
@@ -37,3 +37,7 @@ bindgen = ">=0.68, <0.70"
 tar = "0.4"
 vergen = { version = "8.0.0", features = ["cargo", "git", "gitcl"] }
 walkdir = "2.4"
+
+[features]
+default = []
+gpu-topology = ["dep:nvml-wrapper"]
